Merge with LLK/scratch-flash upstream (#51)
* v453 version bump

* Remove all file extensions when loading a project

When loading a project from a file, remove all instances of '.sb',
'.sb2', or '.sbx' found at the end of the file name. Repeat until none
remain.

Previous behavior: remove either '.sb' or '.sb2' exactly once. Since
this didn't include '.sbx' there are now projects with titles like
'Silly.sbx.sbx.sbx' -- this was the motivation for making this check
repeat.

* Don't list English twice in the language selector

Previous behavior: Always list English first, then append the languages
retrieved from the server. If English is also in that list, then English
will be listed twice.

New behavior: Use the language list retrieved from the server exactly.
If we fail to retrieve a list from the server, use English as a
fallback.

* v453.1 version bump

* HTTP/2 compat: use lowercase header field names

* v454 version bump

* Preserve user text for block args in blockToArray

Before: `blockToArray` returns the parsed value of each block arg
After: `blockToArray` returns the raw text value of each block arg

This means that block arguments should now be preserved by duplicate or
save/load cycles.

* Translation: ask extensions for their specs

This change removes the hard-coded list of extension specs from
`Specs.as`, replacing it with a function on `ExtensionManager` which
reads block specs from the actual extensions. This requires the
extensions to be loaded, so the function pops up a warning dialog if it
detects that any built-in extension is not loaded.

* v455 version bump

* Save block arg internal strings when available

When a block arg's internal storage is a string, save that instead of
the field text. This preserves the internal, universal values (like
`'_mouse_'`) instead of the localized value (like `'mouse-pointer'` or
`'puntero del ratón'`) for selections in drop-down menus.

* Warn about missing extension specs only on export

The new `getExtensionSpecs` method can pop up a warning if one or more
built-in extensions are not loaded. This helps avoid problems when
exporting strings for translation, but we don't want this warning to pop
up when users change their language setting. This change introduces a
Boolean argument to control whether or not the warning should pop up.

* v455.1 version bump

* Revert "Merge pull request #1278 from cwillisf/preserve-menu-items"

This reverts commit 891c6a056871ad79890b4d95e4f295bb50bc16b3, reversing
changes made to 54ece5596f6af9cdeef1bda9f4e02bba283d7d14.

* Revert "Merge pull request #1273 from cwillisf/blockToArray-round-trip-safety"

This reverts commit d0348a0e7f77c07cfe1da24e847daf361bc3980a, reversing
changes made to 23d46beff1050b52a61668bbc616b69d3355f65e.

* v456 version bump

* v456.0.1 version bump

For the 10th anniversary content update

* v456.0.2 version bump

* v456.0.3 version bump

for translation update

* v456.0.4 version bump